add update endpoint and get one tenant endpoint

+++ src/controllers/TenantController.ts
@@ -1,11 +1,15 @@
-import { NextFunction, Response, Request } from 'express';
-import { CreateTenantRequest, ITeanantQuery } from '../types';
-import { inject, injectable } from 'inversify';
-import { TYPES } from '../constants';
-import { ITenantService } from '../services/Interfaces/ITenantService';
-import { ITenantController } from './Interfaces/ITenantConroller';
+import { NextFunction, Request, Response } from 'express';
 import { matchedData, validationResult } from 'express-validator';
+import { inject, injectable } from 'inversify';
+import { AppDataSource } from '../config/data-source';
 import logger from '../config/logger';
+import { TYPES } from '../constants';
+import { Tenant } from '../entity/Tenant';
+import { ITenantService } from '../services/Interfaces/ITenantService';
+import { CreateTenantRequest, ITeanantQuery } from '../types';
+import { ITenantController } from './Interfaces/ITenantConroller';
+import createHttpError from 'http-errors';
+import { log } from 'console';
 
 @injectable()
 class TenantController implements ITenantController {
@@ -18,6 +22,7 @@ class TenantController implements ITenantController {
             res.status(400).json({ errors: result.array() });
             return;
         }
+        logger.debug('Request for creating tenant', { name, address });
         try {
             const tenant = await this.tenantService.create({ name, address });
             logger.info(`Tenant created: ${tenant.id}`);
@@ -30,9 +35,11 @@ class TenantController implements ITenantController {
     async getAll(req: Request, res: Response, next: NextFunction): Promise<void> {
         const validate = matchedData(req, { onlyValidData: true });
 
+        logger.debug('Request for getting all tenants', validate);
         try {
             const [tenants, count] = await this.tenantService.getAll(validate as ITeanantQuery);
             res.json({
+                success: true,
                 currentPage: validate.currentPage as number,
                 perPage: validate.perPage as number,
                 total: count,
@@ -46,10 +53,49 @@ class TenantController implements ITenantController {
 
     async deleteOne(req: Request, res: Response, next: NextFunction): Promise<void> {
         const id = req.params.id;
+
+        if (isNaN(Number(id))) {
+            const err = createHttpError(400, 'Invalid request id');
+            next(err);
+            return;
+        }
+
+        logger.debug('Request for deleting tenant', { id });
         try {
             await this.tenantService.deleteOne(Number(id));
             logger.info('Tenant deleted');
-            res.status(204).json();
+            res.json({ success: true });
+        } catch (error) {
+            next(error);
+        }
+    }
+
+    async updateOne(req: CreateTenantRequest, res: Response, next: NextFunction): Promise<void> {
+        const id = req.params.id;
+        const { name, address } = req.body;
+        logger.debug('Request for updating tenant', { id, name, address });
+        try {
+            const updated = await this.tenantService.updateOne(Number(id), { name, address });
+            logger.info('Tenant updated', updated);
+            res.json({ success: true });
+        } catch (error) {
+            next(error);
+        }
+    }
+
+    async getOne(req: Request, res: Response, next: NextFunction): Promise<void> {
+        const id = req.params.id;
+
+        if (isNaN(Number(id))) {
+            const err = createHttpError(400, 'Invalid request id');
+            next(err);
+            return;
+        }
+        logger.debug('Request for getting a single tenant', { id });
+        try {
+            const tenant = await this.tenantService.getOne(Number(id));
+            logger.info('Tenant fetched', tenant);
+            res.json({ id: tenant.id, name: tenant.name, address: tenant.address });
         } catch (error) {
             next(error);
         }
